Move AuthMixin to user.tests module
